Enable NFC tag presence checking and implement background presence checking.
The background presence checking enables to restart the polling loop if
the application do not release the tag (e.g.: coding or runtime error).
The tag presence will be checked if there is no activity on it during 1s.

The message handler has been moved to the NfcService.java file. It was
previously located in NativeNfcManager,java file for historical reasons.

Change-Id: I3ab5c15c75eb81d2814187b47729859064cf011a
Signed-off-by: Nick Pelly <npelly@google.com>