Use 256-bytes packets in MyTag for better performances.
Also fixed the way the MIU is used. The MIU is the maximum payload
a local socket can receive in a single packet. The MIU may be
different for each side of a socket connection.

When sending data, the application have to make sure the payload
size does NOT exceed the MIU announced by the remote peer for the
current socket.

Change-Id: I3bed94b3042a9383088290de49a81d7ac855c92c