Store the libnfc tag type inside the tag object.
Some functionality in libnfc is different for some technologies
(e.g. reconnect(), transceive()). For these cases, we'll use the libnfc
tag type to deal with them. We don't want to distinguish these cases based on the
Java API technology types, since they may be changed and even mapped to different
libnfc types. Ideally libnfc should abstract this away for us, but that is not
the case now.

Change-Id: I33ea04ca48d16ccf186e3f0882cafdd38a8adb34