Don't call disconnect() twice to avoid emergency recovery.
When a tag is disconnected, first the watchdog is stopped, then disconnect
is called. However, the watchdog stop also did a disconnect() itself.
Now that is only done if the watchdog itself concludes the tag is gone.
Also made sure we catch an early return when the handle is -1;
in that case, just restart the discovery.

Also removed some old code.

Change-Id: Ib851115eecaea65d88cffc7771e273258a404f87