Implement IsoDep timeout handling (NFC service).
Added interface to the NFC service for (re-)setting the IsoDep transaction
timeout. Internally, this is mapped to both the ISO XCHG timer and
the HCI watchdog timer in libnfc (see comments in code).

The timers are reset when the application closes the IsoDep technology.
But they must also be reset if the application crashes (and cannot close
the technology). To cover these cases, whenever discovery is (re)started or
whenever a tag is disconnected, the timeouts are reset to the default values.

Change-Id: I596ef311c3c3b21666b487d16a08642e8a6fc4af