Set LLCP MIU parameter to optimal value.
At the NFCIP-1 level, the LLCP frames are splitted in 251-bytes
packets. Knowing that NFCIP-1 packets have a 3-bytes header, the
optimal size for a LLCP packet is 248-bytes.

With the previous value (256), each LLCP frame was splitted in
two NFCIP-1 packets, one 248+3 bytes and another one of 8+3 bytes.
This was leading to less-than-optimal performances.

Change-Id: Ide31b1ee24efc910c71d690a8ae100331b62ccf8