Fix 4 second latency starting Tag application when recently using home.
The home screen and other system apps call stopAppSwitches() to prevent
malicious applications from starting their activity again when the user exits.
The startActivity() call is delayed by 4 seconds.

We can override this in NFC service because we are a system service, using
resumeAppSwitches(). If a tag comes into the field we really want to show it
immediately. It feels really laggy otherwise.

Change-Id: I5394647f352e3f49f821b5346cd4c9af103137ca