Fix race in NFC tag registry/dispatch.
The tag handle was only registered after the dispatch; with the "normal"
dispatch the activity would take a while to start and this problem would not be
seen. With foreground dispatch however, if you'd immediately try to connect to a tag,
sometimes the tag handle is not registered yet, causing the connect() to fail.

Change-Id: I92b5e36a5d8f08e410a656d364e88bce2d25c854