Check whether an activity is enabled before adding to the chooser.
The RegisteredComponentCache is updated on asynchronous messages
from the PackageManager, which may be delayed for several seconds.
So if an application disables an activity, the activity may still show
up in the activity chooser. Selecting this activity will cause the
NFC service to crash. Prevent this by actually checking whether
an activity is enabled before adding it to the list.

Change-Id: Ibe28b58dd110d93431508cbf728a89743c8fc53b