Animations for 0-click send success / failure.
This implements the following behavior:
- The first part of the animation (the screen scaling down)
  is always played and finished.
- Depending on the result, the following animation will either
  pop the screen back to front (failure), or "clone and swap"
  (success). Both animations will be fully played, except
  if there is a new start() request, in which case they are ended
  immediately.

The main reason to fully play both animations is that the LLCP
transfer typically completes very quickly, and with me profile
removed, none of the animations would ever be visible. I think
the animations are short enough not to be annoying, but we can
tweak depending on user feedback.
Change-Id: Iecd883fec43fad3eeb35f03b6076562b36a10fef