Fix NfcExecutionEnvironment.transceive() failures.
An earlier change to add multiple protocol support for NfcA and IsoDep
causes 3 technologies (NfcA, MifareClassic, IsoDep) to be enumerated on
the Secure Element instead of 2 (MifareClassic, IsoDep). This caused a
regression because
com_android_nfc_jni_open_secure_element_notification_callback() always
connected to the 2nd handle, without checking protocols.

Fix is to walk the protocol list and look for IsoDep instead of hard-coding
the index.

For some reason IsoDep seems to be phNfc_eISO14443_A_PICC and not
phNfc_eISO14443_4A_PICC in this code.

Change-Id: If7471b494099a09cf6d0bf5e255da556f11570b1