NFC: Remove statusbar/navbar from animated screenshot.
This just crops the status bar and the navbar from the screenshot we take.
This makes the animation look right in all cases, except for when we are full screen:
in that scenario we're cropping away too much.

Also removed the clone view (was no longer used) and cleaned FireflyRenderer
a bit.

Bug: 5318512
Change-Id: I37189a576f19937c2cae4ec368f26ac0725a1b62