Cache installed packages for sending SE broadcasts.
Getting the list of installed packages takes some time
on a phone with lots of packages (seen >500ms). Since
we did this for every SE broadcast (including field on / field off),
I've seen the NFC service event queue fill up, which has
a bad effect on latency-critical events such as the events
for llcp link establishment.

This caches the list of installed packages.

Bug: 5963061
Change-Id: Ia9f02c9ce58dae5c81949f7317425e5ee415ee92