Don't close the NFC-EE pipe immediately on disable.
The combination of an outstanding APDU request together
with a "warm-reset" HCI command that we send on NFC disable
sometimes cause a complete NFC controller lockup (depending
on the order and timing).

To prevent this, delay closing the NFC-EE pipe for a while on
NFC disable to give processes a chance to finish their outstanding
APDU request.

Bug: 8660403
Change-Id: I40d5f33fc41453aa3592a50721b8bef793e0af5d