Support chip-specific max EE setting and configuring 0 EE.
Added 'get_max_ee' to the HAL interface to allow the NFC stack to
inquire the HAL layer of the maximum number of EE on the platform.

Added the necessary HAL interface to allow the stack to get the maximum
number of EE on the system.  This allows the system to dynamically set
this value at run-time in order to support chip-specific differences
in the number of EE on the platform.

Change-Id: I3ff00d117d715b1f46636501932ff92accc39483