HCE: Stop routing AIDs for non-default payment services.
The logic implemented in KitKat would still route AIDs
for non-default payment services:
1) If there was no AID overlap at all, it would route
   AIDs and not ask for any confirmation.
2) If there was AID overlap, it would route AID but
   ask the user to confirm to deviate from the default.

In practice this logic is tricky in the transition
scenario we're in, where existing secure-element
based apps may not register their AIDs. This would
cause the conflict-detection mechanism to fail, and
inconsistent behavior to occur.

With this change, only the AIDs of the default selected
payment service are routed. All non-default payment
services won't have their AIDs routed, independent of whether
they are conflicting with the default or not.

Change-Id: Ibc588d7c81f0478a71021064bd63f5f6c91bd2fe