Adds functionality to detect live case NFC tags on lock screen.
Live case usecase at lock screen can be disabled by setting
enable_live_cases in the live_cases.xml to false.
live_case_tag_types contains technology types which will be polled for
at the lock screen. And live_case_mime_types contains the list of mime
records that will be allowed to be passed to the application. Currently
they are set to the Live Case application mime type, and can be
overriden for each of the device.

Test: Manual; test with live case
Change-Id: I2534aeeaf550dee128c8963dfbada40574d5a4a5
Signed-off-by: Ruchi Kandoi<kandoiruchi@google.com>