Use checksum to determine if config has been modified
Using a timestamp is not reliable as some partitions on Android reset
the modification timestamp in certain partitions on all files. This
leads to the NFC chip configuration not being loaded in some instances,
even so it has been modified.

This patch switches the file modification check from a timestamp to
calculate a simple CRC32 checksum instead.

Bug: 67959484
Test: manual; Verify configuration is loaded when modified
Change-Id: I57723dd5d2e10813806f492551d7ed502124a295