Add nfc_utils_ringbuffer_fuzzer
This is a basic fuzz target for the ringbuffer implementation for NFC.
This is shared by bluetooth. It looks as if the only users are debug
snoopers for NFC and Bluetooth.

This uncovers a bug that doesn't currently impact the platform. There's
a bug wherein a ringbuffer initialised with ringbuffer_init(0) (size ==
0 bytes) manifests in a FPE when ringbuffer_peek/ringbuffer_pokp is
called on it. This bug can be triggered by the corpus entry.

This is a real bug, but all users (it looks like) currently have a
macro-defined size that is nonzero. Worth fixing the bug though.

Bug: N/A
Test: m nfc_utils_ringbuffer_fuzzer
Change-Id: I907ace09eb0c1aff52799f5b4b4d17863f548efc