NFC: Fix missing mServiceBound state update
Failure to update this variable will cause mService and mServiceName not
to be nullified if onHostEmulationDeactivated() is called right after
onServiceConnected() is called. onHostEmulationDeactivated() calls
unbindServiceIfNeededLocked() which uses mServiceBound to unbind the
service. Since mServiceBound was not updated to the correct value,
unbindServiceIfNeededLocked() will never be able to unbind the
previously bound service.

Bug: 178754309
Test: Service rebinds properly in the following sequence: New service is
bound -> onHostEmulationDeactivated() is called -> New service is seen
to bind again when onHostEmulationActivated() is called.

Signed-off-by: Shiv Rajora <srajora@google.com>
Change-Id: Ibb75c5ca0a7dfd25ee85ed770fed7cc24caa7ab0