use test app's data folder as filter directory
During a retry, test methods to be retried are written into a
file, pushed onto device and read by the test instrumentation
for execution.

The default location for such filter file is /data/local/tmp,
which is prohibited for access by system uid packages, such
as the system UI tests package.

This change uses the test app's data folder for filter: this
is not generally applicable for all the tests since root
permission is required, but it is a relatively safe assumption
in case of testing against system uid pakcages since it
implies userdebug build is being used.

Bug: 152001555
Change-Id: I510f75100f81fabe1b7addffdb2840597f30c841
Test: http://go/forrest-run/L49600000952176618